Finished enrollment lifecycle.
The process of enrollment can now be completed by explicitly POSTing to /api/enrollment/completion with a request parameter of withoutEncounters set to true or false. In the latter case, we verify that at least one encounter has been registered for the tracked person. An enrollment can be reopened by issuing a DELETE request to /api/enrollment/completion. Those state transitions are now also communicated in the enrollment resource itself. Responses of the individual state transitions now also point back to the enrollment resource in their Location headers.

Added unit tests for the state transitions in TrackedCase.