Make sure that domain object reference resolution errors make it into the Errors, too.
The aggregate resolution customization deployed into the ModelMapper now throws a dedicated exception in case it can't find a target aggregate for an identifier submitted. We also now use a MapperWrapper (haha!) to handle those exceptions and propagate them into an Either<T, Errors> that can then be ….fold(…)-ed into a response.

Usage of MapperWrapper is now preferred as it also returns the target object as is for calls to …map(…) for a destination object.